PREAPPS-558: Move selection logic out of `GuiRichTextArea` and `MiniComposer` and into base `RichTextArea`.
